Move rendered methods to abstract base class
