+++ src/main/java/com/elovirta/dita/markdown/renderer/AbstractRenderer.java
@@ -8,15 +8,35 @@ import static org.dita.dost.util.Constants.*;
 import static org.dita.dost.util.URLUtils.toURI;
 import static org.dita.dost.util.XMLUtils.AttributesBuilder;
 
+import com.elovirta.dita.markdown.DitaRenderer;
 import com.elovirta.dita.markdown.SaxWriter;
+import com.elovirta.dita.utils.ClasspathURIResolver;
+import com.google.common.base.Suppliers;
 import com.google.common.io.Files;
-import com.vladsch.flexmark.ast.Heading;
-import com.vladsch.flexmark.ast.Text;
+import com.vladsch.flexmark.ast.*;
+import com.vladsch.flexmark.ext.anchorlink.AnchorLink;
+import com.vladsch.flexmark.ext.attributes.AttributesNode;
+import com.vladsch.flexmark.ext.gfm.strikethrough.Strikethrough;
+import com.vladsch.flexmark.ext.gfm.strikethrough.Subscript;
+import com.vladsch.flexmark.ext.superscript.Superscript;
 import com.vladsch.flexmark.util.ast.Document;
 import com.vladsch.flexmark.util.ast.Node;
+import com.vladsch.flexmark.util.data.DataHolder;
+import com.vladsch.flexmark.util.sequence.BasedSequence;
+import java.io.IOException;
+import java.io.InputStream;
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import javax.xml.transform.Templates;
+import javax.xml.transform.TransformerConfigurationException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.sax.SAXTransformerFactory;
+import javax.xml.transform.stream.StreamSource;
 import org.dita.dost.util.DitaClass;
 import org.xml.sax.Attributes;
 
@@ -37,10 +57,112 @@ public abstract class AbstractRenderer {
   protected static final Attributes ALT_ATTS = buildAtts(TOPIC_ALT);
   protected static final Attributes PH_ATTS = buildAtts(TOPIC_PH);
 
+  protected final boolean mditaExtendedProfile;
+  protected final boolean mditaCoreProfile;
+  protected final Supplier<SAXTransformerFactory> transformerFactorySupplier;
+  protected final Supplier<Templates> templatesSupplier;
+
+  public AbstractRenderer(DataHolder options) {
+    mditaExtendedProfile = DitaRenderer.MDITA_EXTENDED_PROFILE.getFrom(options);
+    mditaCoreProfile = DitaRenderer.MDITA_CORE_PROFILE.getFrom(options);
+    transformerFactorySupplier =
+      Suppliers.memoize(() -> {
+        final SAXTransformerFactory tf = (SAXTransformerFactory) TransformerFactory.newInstance();
+        tf.setURIResolver(new ClasspathURIResolver(tf.getURIResolver()));
+        return tf;
+      })::get;
+    templatesSupplier =
+      Suppliers.memoize(() -> {
+        final SAXTransformerFactory tf = transformerFactorySupplier.get();
+        final String stylesheet = (mditaCoreProfile || mditaExtendedProfile)
+          ? "/hdita2dita.xsl"
+          : "/hdita2dita-markdown.xsl";
+        try (InputStream in = getClass().getResourceAsStream(stylesheet)) {
+          return tf.newTemplates(new StreamSource(in, "classpath://" + stylesheet));
+        } catch (IOException | TransformerConfigurationException e) {
+          throw new RuntimeException(e);
+        }
+      })::get;
+  }
+
   /**
    * @return the mapping of nodes this renderer handles to rendering function
    */
-  abstract Map<Class<? extends Node>, NodeRenderingHandler<? extends Node>> getNodeRenderingHandlers();
+  protected Map<Class<? extends Node>, NodeRenderingHandler<? extends Node>> getNodeRenderingHandlers() {
+    final List<NodeRenderingHandler> res = new ArrayList<>();
+    res.add(new NodeRenderingHandler<>(AnchorLink.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(Code.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(TextBase.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(Emphasis.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(StrongEmphasis.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(HtmlEntity.class, (node, context, html) -> render(node, context, html)));
+    res.add(new NodeRenderingHandler<>(HtmlInlineComment.class, (node, context, html) -> render(node, context, html)));
+    if (!mditaCoreProfile) {
+      res.add(new NodeRenderingHandler<>(Superscript.class, (node, context, html) -> render(node, context, html)));
+      res.add(new NodeRenderingHandler<>(Subscript.class, (node, context, html) -> render(node, context, html)));
+    }
+    if (!mditaCoreProfile && !mditaExtendedProfile) {
+      res.add(new NodeRenderingHandler<>(Strikethrough.class, (node, context, html) -> render(node, context, html)));
+    }
+
+    return res.stream().collect(Collectors.toMap(handler -> handler.getNodeType(), handler -> handler));
+  }
+
+  protected void render(AnchorLink node, final NodeRendererContext context, final SaxWriter html) {
+    context.renderChildren(node);
+  }
+
+  private void render(final Code node, final NodeRendererContext context, final SaxWriter html) {
+    if (mditaExtendedProfile) {
+      printTag(node, context, html, HI_D_TT, TT_ATTS);
+    } else {
+      printTag(node, context, html, PR_D_CODEPH, getInlineAttributes(node, CODEPH_ATTS));
+    }
+  }
+
+  protected void render(final TextBase node, final NodeRendererContext context, final SaxWriter html) {
+    context.renderChildren(node);
+  }
+
+  protected void render(final Emphasis node, final NodeRendererContext context, final SaxWriter html) {
+    printTag(node, context, html, HI_D_I, getInlineAttributes(node, I_ATTS));
+  }
+
+  protected void render(final StrongEmphasis node, final NodeRendererContext context, final SaxWriter html) {
+    printTag(node, context, html, HI_D_B, getInlineAttributes(node, B_ATTS));
+  }
+
+  protected void render(final Superscript node, final NodeRendererContext context, final SaxWriter html) {
+    printTag(node, context, html, HI_D_SUP, getInlineAttributes(node, SUP_ATTS));
+  }
+
+  protected void render(final Subscript node, final NodeRendererContext context, final SaxWriter html) {
+    printTag(node, context, html, HI_D_SUB, getInlineAttributes(node, SUB_ATTS));
+  }
+
+  private void render(final Strikethrough node, final NodeRendererContext context, final SaxWriter html) {
+    if (mditaExtendedProfile) {
+      printTag(node, context, html, TOPIC_PH, PH_ATTS);
+    } else {
+      printTag(node, context, html, HI_D_LINE_THROUGH, getInlineAttributes(node, LINE_THROUGH_ATTS));
+    }
+  }
+
+  /**
+   * Map HTML entity to Unicode character.
+   */
+  private void render(final HtmlEntity node, final NodeRendererContext context, final SaxWriter html) {
+    final BasedSequence chars = node.getChars();
+    final String name = chars.subSequence(1, chars.length() - 1).toString().toLowerCase();
+    final String val = Entities.ENTITIES.getProperty(name);
+    if (val != null) {
+      html.characters(val);
+    }
+  }
+
+  private void render(final HtmlInlineComment node, final NodeRendererContext context, final SaxWriter html) {
+    // Ignore
+  }
 
   protected boolean hasMultipleTopLevelHeaders(Document astRoot) {
     final long count = StreamSupport
@@ -133,5 +255,31 @@ public abstract class AbstractRenderer {
       sb.append(Character.toLowerCase(c));
     }
     return sb.toString();
+  }
+
+  protected Attributes getInlineAttributes(Node node, Attributes base) {
+    if (!mditaExtendedProfile) {
+      if (node.getChildOfType(AttributesNode.class) != null) {
+        final Title header = Title.getFromChildren(node);
+        final AttributesBuilder builder = new AttributesBuilder(base);
+        return readAttributes(header, builder).build();
+      } else if (node.getNext() instanceof AttributesNode) {
+        final Title header = Title.getFromNext(node);
+        final AttributesBuilder builder = new AttributesBuilder(base);
+        return readAttributes(header, builder).build();
+      }
+    }
+    return base;
+  }
+
+  protected AttributesBuilder readAttributes(Title header, AttributesBuilder builder) {
+    if (!header.classes.isEmpty()) {
+      builder.add(ATTRIBUTE_NAME_OUTPUTCLASS, String.join(" ", header.classes));
+    }
+    for (Map.Entry<String, String> attr : header.attributes.entrySet()) {
+      builder.add(attr.getKey(), attr.getValue());
+    }
+    header.id.ifPresent(id -> builder.add(ATTRIBUTE_NAME_ID, id));
+    return builder;
   }
 }
